# swagger.yml
openapi: 3.0.0
info:
  description: "The REST API for my demo e-commerce app, based on various herb and spice online shopping platforms."
  version: "1.0.0"
  title: "Mikayla's Spice Shop, REST API"
schemes: 
  - http
servers:
- url: localhost:8088/api
  description: The URL which currently hosts the REST API
host: localhost:4000
basePath: "/"
...
paths:
...
  /regions:
    get:
      summary: "Get all product regions"
      tags:
        - region
      responses:
        200:
          description: "Got region listing successfully"
          schema:
            type: object
        404:
          description: "Requested resource was not found"
          schema:
            type: object
  /regions/{regionid}:
    get:
      summary: "Get one region by its ID"
      tags:
        - region
      responses:
        200:
          description: "Got region listing successfully"
          schema:
            type: object
        404:
          description: "Requested resource was not found"
          schema:
            type: object
...